% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpLabel.R
\name{cpvisLabel}
\alias{cpvisLabel}
\alias{cpLabel}
\title{Visual Interface for Changepoint Penalty Exploration (D3 version)}
\usage{
cpLabel(data)
}
\arguments{
\item{data}{A univariate dataset}
}
\value{
starts a shiny app in a new window
}
\description{
This function takes a univariate dataset as input. Provides simple labeling and exporting functionality for univariate time seriese data.
}
\examples{
data = c(rnorm(100,0,1),rnorm(100,5,1))
cpLabel(data)


}
