% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpvis_d3.R
\name{cpVisualise}
\alias{cpVisualise}
\title{Visual Interface for Unsupervised Changepoint Penalty Exploration}
\usage{
cpVisualise(data, penalty_range = c(1e-05, 10))
}
\arguments{
\item{data}{A univariate dataset}

\item{penalty_range}{A range of penalty values}
}
\value{
starts a shiny app in a new window
}
\description{
This function takes a univariate dataset as input, as well as a range of penalty values.
It uses cpt.mean with the "PELT" method and "CROPS" penalty type from the changepoint package.
The function opens a web interface that displays:
\enumerate{
\item The univariate datset (top left)
\item General statistics (top right)
\item Solution path (bottom left)
\item Weighted means histogram of the segments between changepoints.
}
}
\details{
The main form of interaction with the interface is through selecting each penalty value and then inspecting the resulting solution on the main plot (top left).
}
\examples{
data = c(rnorm(100,0,1),rnorm(100,5,1))
cpVisualise(data, penalty_range = c(1e-5,10))
}
