% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpvis_d3.R
\name{cpvisVisualise}
\alias{cpvisVisualise}
\alias{cpVisualise}
\title{Visual Interface for Changepoint Penalty Exploration (D3 version)}
\usage{
cpVisualise(data, penalty_range = c(1e-05, 10))
}
\arguments{
\item{data}{A univariate dataset}

\item{penalty_range}{A range of penalty values}
}
\value{
starts a shiny app in a new window
}
\description{
This function takes a univariate dataset as input, as well as a range of penalty values.
It uses cpt.mean with the "PELT" method and "CROPS" penalty type from the changepoint package.
}
\examples{
data = c(rnorm(100,0,1),rnorm(100,5,1))
cpVis_d3(data, penalty_range = c(1e-5,10))

}
