% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpvis_d3.R
\name{cpVisualise}
\alias{cpVisualise}
\title{Visual Interface for Unsupervised Changepoint Penalty Exploration}
\usage{
cpVisualise(data, penalty_range = c(1e-5,10))
}
\arguments{
\item{data}{A univariate dataset. This should be a vector containing the data within which you wish to find a changepoint.}

\item{penalty_range}{A range of penalty values. This will vary based on the datset and is used to compute changepoints for the data for each penalty value.}
}
\value{
Starts a shiny app in a new window.

Saves the currently selected solution as a .csv file, for use in cpLabel.
}
\description{
{ This function calculates the optimal positioning and (potentially) number of changepoints for a given univariate dataset using the user
specified range of penalty values.
It uses cpt.mean with the "PELT" method and "CROPS" penalty type from the changepoint package.


The function starts a shiny server and acompanying web interface that displays:
\enumerate{
\item The univariate datset (top left)
\item General statistics (top right)
\item Solution path (bottom left)
\item Weighted means histogram of the segments between changepoints.
}
The main form of interaction with the interface is through selecting each penalty value and then inspecting the resulting solution on the main plot (top left).

}
}
\details{
{This function is used to find changes in mean for a given dataset using the PELT method.
The function visualises the data, along side an interactive solution path, allowing the usert to inspect how different penalty values impact the resulting changepoint locations.
Additional information is also presented to help the user interpret the data and specific penalty value. A table containing general statistics such as mean, variance etc and a weighted histogram, showing the distribution of segment means (weighted by length) for a particular penalty value.

For an extended usage example, please see the accompanying website: \url{https://sites.google.com/view/changepointvis/home}

}
}
\examples{
\dontrun{
data = c(rnorm(100,0,1),rnorm(100,5,1))
penalty_range = c(1e-5,10)
cpVisualise(data, penalty_range)
}

}
\references{
{
PELT Algorithm: Killick R, Fearnhead P, Eckley IA (2012) Optimal detection of changepoints with a linear computational cost, JASA 107(500), 1590–1598 \cr
CROPS: Haynes K, Eckley IA, Fearnhead P (2014) Efficient penalty search for multiple changepoint problems (in submission), arXiv:1412.3617}
}
\author{
Oliver Ford
}
